Build python3 packages
Build python3-* versions of autopilot, -tests, -trace, and -vis. Don't update
the metapackages yet, as we want to avoid pulling in the python3 stack into the
touch images. We might do separate -py3 metapackages at some point, or just
switch over the current ones to -py3.

Note, this does NOT copy the postrm script to remove the autopilot group. Doing
that is generally discouraged, and it causes problems if you remove autopilot
for one Python version but keep the package for the other.

Respect DEB_BUILD_OPTIONS=nocheck.