bamf: Fix initialization of GdkDisplay
GdkX11.X11Display has no public constructor, so it's invalid to create a
GdkDisplay like that. The official API is to use Gdk.Display.get_default(), use
that instead. This fixes a crash when trying to access the "geometry" Window
property.

Add a test case for Window.geometry plausibility, as this code currently does
not have any tests. (This reproduces the crash.)