Don't try to call nonexisting bin/autopilot
The bin/autopilot wrapper disappeared recently. Update the functional test
runner and the docs to invoke the tests with "$PYTHON -m autopilot.run"
instead.

This fixes running the functional tests, as otherwise they would always use the
installed "autopilot" program (which may not exist at all, or be too old).