Fix the autopilot-touch package to be Architecture: all instead of having it be Architecture: any with a wrong architecture exclusion of a dependency.
It appears the only reason autopilot-touch is currently Architecture: any
instead of Architecture: all is in order to have different dependencies on
powerpc.  But that makes the dependencies of the package on powerpc *wrong*;
the autopilot-touch package without libautopilot-qt is incomplete.

The right fix here is either to ensure autopilot-touch is not built at all
on powerpc (or indeed, on all architectures other than amd64 i386 armhf), or
to just make it an Architecture: all package and allow it to be
uninstallable on powerpc (which is perfectly allowable in the archive).